requirements.txt
click==8.0.3
mammoth==1.5.0
python-docx==0.8.11
PyYAML==6.0
